backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai>=0.3.1,<0.4.0
langchain==0.0.340
chromadb==0.4.18
python-multipart==0.0.6
markdown==3.5.1
pydantic==2.5.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
langchain-google-genai==0.0.6
